% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dzmpl}
\alias{dzmpl}
\title{Zero Modified Poisson-Lindley (ZMPL) distribution}
\usage{
dzmpl(x, theta,  p0 = 0, log = FALSE)
pzmpl(q, theta = 5, p0 = 0, lower.tail = TRUE, log.p = FALSE)
qzmpl(p, theta = 5, p0 = 0, lower.tail = TRUE, log.p = FALSE)
rzmpl(n, theta = 5, p0 = 0)
}
\arguments{
\item{x, q}{vector of observations/quantiles}

\item{theta}{vector of scale parameter values}

\item{p0}{vector of shape parameter values}

\item{log, }{log.p logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}
}
\value{
returns a object of the ZMPL distribution.
}
\description{
Density, distribution function, quantile function and random generation for the ZMPL distribution.
}
\details{
If \code{theta} or \code{p0} not specified they assume the default values of 5 and 0, respectively.

The ZMPL distribution has density:

\deqn{
Pr(X=k)=\left\{\begin{array}{ll} \pi + (1 - \pi)\frac{\theta^2(\theta + 2)}{(\theta + 1)^3},&k = 0,  \\&\\(1 - \pi)\frac{\theta^2(k + \theta + 2)}{(\theta + 1)^{k+3}},&k \in \mathbb{N}^{ {\tiny +}}. \end{array}\right.
}
}
\note{
For the function ZMPL(), p0 is the zero-modification
parameter and different values lead to different
modifications of the ZMPL distribution.
}
\examples{
dat <- rzmpl(1000); plot(table(dat),type="h",xlab="x",ylab="PMF")
fit <- fitzmpl(dat)

}
\author{
Manoel Santos-Neto \email{mn.neco@gmail.com}, Danillo Xavier \email{danilloxavier@gmail.com}, Marcelo Bourguignon \email{m.p.bourguignon@gmail.com} and Vera Tomazella \email{vera@ufscar.com}
}
\references{
Santos-Neto, M., Xavier, D., Bourguignon, M. and Tomazella, V. (2017) Zero-Modified Poisson-Lindley distribution with applications in inflated and deflated of zeros.
\emph{to appear.}
}

