% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fitted.zmpl}
\alias{fitted.zmpl}
\title{Zero Modified Poisson-Lindley (ZMPL) distribution}
\usage{
dZMPL(x, mu = 1, sigma = 1, nu=0.1, log = FALSE)
pZMPL(q, mu = 1, sigma = 1,  nu=0.1, lower.tail = TRUE, log.p = FALSE)
qZMPL(p, mu = 1, sigma = 1,  nu=0.1, lower.tail = TRUE, log.p = FALSE)
rZMPL(n, mu = 1, sigma = 1)
plotZMPL(mu = .5, sigma = 1,  nu=0.1, from = 0, to = 0.999, n = 101, ...)
meanZMPL(obj)
}
\arguments{
\item{x, q}{vector of observations/quantiles}

\item{mu}{vector of scale parameter values}

\item{sigma}{vector of shape parameter values}

\item{log, }{log.p logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}

\item{from}{where to start plotting the distribution from}

\item{to}{up to where to plot the distribution}

\item{...}{other graphical parameters for plotting}
}
\value{
returns a object of the ZMPL distribution.
}
\description{
The fuction \code{ZMPL()} defines the ZMPL distribution, a two paramenter
distribution. The zero modified Poisson-Lindley distribution is similar
to the Poisson-Lindley distribution but allows zeros as y values. The extra parameter
models the probabilities at zero. The functions dZMPL, pZMPL, qZMPL and rZMPL define
the density, distribution function, quantile function and random generation for
the zero modified Poisson-Lindley distribution.
plotZMPL can be used to plot the distribution. meanZMPL calculates the expected
value of the response for a fitted model.
}
\details{
The probability massa function of the is given by

\deqn{ f_{Y}(y;\mu,\delta,p) =\frac{[1-p]\sqrt{\delta+1}}{4\,y^{3/2}\,\sqrt{\pi\mu}}\left[y+\frac{\delta\mu}{\delta+1} \right]\exp\left(-\frac{\delta}{4}\left[\frac{y[\delta+1]}{\delta\mu}+\frac{\delta\mu}{y[\delta+1]}-2\right]\right) I_{(0, \infty)}(y)+  pI_{\{0\}}(y).}
}
\note{
For the function ZMPL(), mu is the mean and sigma is the precision parameter and nu is the proportion of zeros of the zero adjusted Birnbaum-Saunders distribution.
}
\examples{
plotZARBS()
dat <- rZARBS(1000); hist(dat)
fit <- gamlss(dat~1,family=ZARBS(),method=CG())
meanZABS(fit)

data(papatoes);
fit = gamlss(I(Demand/10000)~1,sigma.formula=~1, nu.formula=~1, family=ZARBS(mu.link="identity",sigma.link = "identity",nu.link = "identity"),method=CG(),data=papatoes)
summary(fit)

data(oil)
fit1 = gamlss(Demand~1,sigma.formula=~1, nu.formula=~1, family=ZARBS(mu.link="identity",sigma.link = "identity",nu.link = "identity"),method=CG(),data=oil)
summary(fit1)

}
\author{
Manoel Santos-Neto \email{manoel.ferreira@ufcg.edu.br}, F.J.A. Cysneiros \email{cysneiros@de.ufpe.br}, Victor Leiva \email{victorleivasanchez@gmail.com} and Michelli Barros \email{michelli.karinne@gmail.com}
}
\references{
Leiva, V., Santos-Neto, M., Cysneiros, F.J.A., Barros, M. (2015) A methodology for stochastic inventory models based on a zero-adjusted Birnbaum-Saunders distribution.
\emph{Applied Stochastic Models in Business and Industry.} \email{10.1002/asmb.2124}.
}

