% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{FIZMPL}
\alias{FIZMPL}
\title{Zero Modified Poisson-Lindley (ZMPL) distribution}
\usage{
dZMPL(x, theta = 1, p0=0.1, log = FALSE)
pZMPL(q, theta = 1, p0=0.1, lower.tail = TRUE, log.p = FALSE)
qZMPL(p, theta = 1, p0=0.1, lower.tail = TRUE, log.p = FALSE)
rZMPL(n, theta = 1, p0=0.1)
plotZMPL(theta = 1, p0=0.1, from = 0, to = 0.999, n = 101, ...)
meanZMPL(obj)
}
\arguments{
\item{theta}{theta parameter values}

\item{p0}{p0 parameter values}

\item{x, q}{vector of observations/quantiles}

\item{log, }{log.p logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}

\item{from}{where to start plotting the distribution from}

\item{to}{up to where to plot the distribution}

\item{...}{other graphical parameters for plotting}
}
\description{
The fuction \code{ZMPL()} defines the ZMPL distribution, a two paramenter
distribution. The zero modified Poisson-Lindley distribution is similar
to the Poisson-Lindley distribution but allows zeros as y values. The extra parameter
models the probabilities at zero. The functions dZMPL, pZMPL, qZMPL and rZMPL define
the density, distribution function, quantile function and random generation for
the zero modified Poisson-Lindley distribution.
plotZMPL can be used to plot the distribution. meanZMPL calculates the expected
value of the response for a fitted model.
}
\details{
The probability massa function of the is given by

\deqn{ f_{X}(x;\theta,p0) = p0 + (1-p0)*P[Y=0] I(y=0) + (1-p0)*P[Y>0] I(y>0)}
where \deqn{P[Y=y]} is a probability massa function of the Poisson-Lindley distribution.
}
\examples{
FIZMPL()
theta <- seq(0.05,1,l=100)
aux0 <- - ((theta^2)*(theta+2)/( (theta^2) +3*theta+1) )
p0 <- seq(0,1,l=100)
g <- expand.grid(x = theta, y = p0)
z <- FIZMPL(theta,p0)
z[z<0] <- 0
g$z <- z
print(wireframe(z ~ x * y  ,g, xlab=expression(theta), scales = list(arrows = FALSE), ylab=expression(pi), zlab = "FI(X)") )


theta <- seq(0.05,1,l=100)
aux0 <- - ((theta^2)*(theta+2)/( (theta^2) +3*theta+1) )
p0 <- seq(min(aux0),0,l=100)
g <- expand.grid(x = theta, y = p0)
z <- FIZMPL(theta,p0)
z[z<0] <- 0
g$z <- z
print(wireframe(z ~ x * y  ,g, xlab=expression(theta), scales = list(arrows = FALSE), ylab=expression(pi), zlab = "FI(X)") )
}
\author{
Manoel Santos-Neto \email{mn.neco@gmail.com}, Marcelo Bourguignon \email{m.p.bourguignon@gmail.com}, Danillo Xavier \email{danilloxavier@gmail.com} and Vera Tomazella \email{vera@ufscar.br}
}
\references{
Bourguignon, M., Xavier, D., Santos-Neto, M., Tomazella, V. (2017) The Modified Poisson-Lindley distribution: A model for overdispersed and underdispersed count data.
\emph{manuscript.}
}

