% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesisTesting.R
\name{grad.test}
\alias{grad.test}
\title{Dispersion test}
\usage{
grad.test(modelh0,modelh1)
}
\arguments{
\item{modelh0}{model under null hypothesis.}

\item{modelh1}{model under alternative hypothesis.}
}
\value{
A list with class "htest" containing the following components:

\code{statistic}	the value of the test statistic.

\code{parameter}	the degrees of freedom for the test statistic.

\code{p.value}	the p-value for the test.

\code{method}	a character string indicating what type of likelihood ratio test was performed.

\code{data.name} a character string giving the name(s) of the data
}
\description{
Tests the null hypothesis of precision fixed in RBS models against the alternative of precision variable.
}
\examples{

##
library(alr3)
data(landrent)
attach(landrent)
resp <- I(Y/X1)
y1 <-  split(resp, X4)$"1"
x21 <-  split(X2, X4)$"1"
##Fixed Precision
fit0 <- gamlss(y1 ~ x21, family=RBS(mu.link="identity"),method=CG()  )
##Varying Precision
fit1 <- gamlss(y1 ~ x21,sigma.formula = y1 ~x21, family=RBS(mu.link="identity",sigma.link="sqrt"),method=CG()  )
#Precision Test
lr.test(fit0,fit1)
score.test(fit0,fit1)
grad.test(fit0,fit1)
wald.test(fit1)
}
\author{
Manoel Santos-Neto \email{manoel.ferreira@ufcg.edu.br}, F.J.A. Cysneiros \email{cysneiros@de.ufpe.br}, Victor Leiva \email{victorleivasanchez@gmail.com} and Michelli Barros \email{michelli.karinne@gmail.com}
}

