% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qzmpl}
\alias{qzmpl}
\title{Zero Modified Poisson-Lindley (ZMPL) distribution}
\usage{
dzmpl(x, theta,  p0 = 0, log = FALSE)
pzmpl(q, theta = 5, p0 = 0, lower.tail = TRUE, log.p = FALSE)
qzmpl(p, theta = 5, p0 = 0, lower.tail = TRUE, log.p = FALSE)
rzmpl(n, theta = 5, p0 = 0)
}
\arguments{
\item{p}{vector of probabilities.}

\item{theta}{vector of scale parameter values}

\item{p0}{vector of shape parameter values}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]}

\item{x, q}{vector of observations/quantiles}

\item{log, }{log.p logical; if TRUE, probabilities p are given as log(p).}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}
}
\value{
returns a object of the ZMPL distribution.
}
\description{
Density, distribution function, quantile function and random generation for the ZMPL distribution.
}
\details{
If \code{theta} or \code{p0} not specified they assume the default values of 5 and 0, respectively.

The ZMPL distribution has density:S

\deqn{ f_{Y}(y;\mu,\delta,p) =\frac{[1-p]\sqrt{\delta+1}}{4\,y^{3/2}\,\sqrt{\pi\mu}}\left[y+\frac{\delta\mu}{\delta+1} \right]\exp\left(-\frac{\delta}{4}\left[\frac{y[\delta+1]}{\delta\mu}+\frac{\delta\mu}{y[\delta+1]}-2\right]\right) I_{(0, \infty)}(y)+  pI_{\{0\}}(y).}
}
\note{
For the function ZMPL(), mu is the mean and sigma is the precision parameter and nu is the proportion of zeros of the zero adjusted Birnbaum-Saunders distribution.
}
\examples{
plotZARBS()
dat <- rZARBS(1000); hist(dat)
fit <- gamlss(dat~1,family=ZARBS(),method=CG())
meanZABS(fit)

data(papatoes);
fit = gamlss(I(Demand/10000)~1,sigma.formula=~1, nu.formula=~1, family=ZARBS(mu.link="identity",sigma.link = "identity",nu.link = "identity"),method=CG(),data=papatoes)
summary(fit)

data(oil)
fit1 = gamlss(Demand~1,sigma.formula=~1, nu.formula=~1, family=ZARBS(mu.link="identity",sigma.link = "identity",nu.link = "identity"),method=CG(),data=oil)
summary(fit1)

}
\author{
Manoel Santos-Neto \email{manoel.ferreira@ufcg.edu.br}, F.J.A. Cysneiros \email{cysneiros@de.ufpe.br}, Victor Leiva \email{victorleivasanchez@gmail.com} and Michelli Barros \email{michelli.karinne@gmail.com}
}
\references{
Leiva, V., Santos-Neto, M., Cysneiros, F.J.A., Barros, M. (2015) A methodology for stochastic inventory models based on a zero-adjusted Birnbaum-Saunders distribution.
\emph{Applied Stochastic Models in Business and Industry.} \email{10.1002/asmb.2124}.
}

